Refactor Functions to internal package (EL only)
