+++ src/main/resources/META-INF/services/org.kaazing.k3po.lang.el.spi.FunctionMapperSpi
@@ -1,1 +1,1 @@
-org.kaazing.specification.ws.Functions$Mapper
+org.kaazing.specification.ws.internal.Functions$Mapper
